Fix flags definition

+++ flag.go
@@ -39,19 +39,19 @@ func FlagStringSlice(name string, shorthand string, dest *[]string, description 
 	return &flag{name: name, shorthand: shorthand, description: description, destination: dest}
 }
 
-func FlagInt(name string, shorthand string, dest *int8, description string) Flag {
+func FlagInt(name string, shorthand string, dest *int, description string) Flag {
 	return &flag{name: name, shorthand: shorthand, description: description, destination: dest}
 }
 
-func FlagIntSlice(name string, shorthand string, dest *[]int8, description string) Flag {
+func FlagIntSlice(name string, shorthand string, dest *[]int, description string) Flag {
 	return &flag{name: name, shorthand: shorthand, description: description, destination: dest}
 }
 
-func FlagUint(name string, shorthand string, dest *int8, description string) Flag {
+func FlagUint(name string, shorthand string, dest *uint, description string) Flag {
 	return &flag{name: name, shorthand: shorthand, description: description, destination: dest}
 }
 
-func FlagUIntSlice(name string, shorthand string, dest *[]int16, description string) Flag {
+func FlagUIntSlice(name string, shorthand string, dest *[]uint, description string) Flag {
 	return &flag{name: name, shorthand: shorthand, description: description, destination: dest}
 }
